Extract lifecycle bootstrap onto Identity and regroup it with Setup
Fold setup's identity construction, handler install, archive flush,
stale-record recovery and hub triggers into `Identity.bootstrapLifecycle()`,
and move the foreground/background wiring into `Identity.table()`. `setup`
now reads as top-level orchestration: build the identity, bootstrap it, start
listening, bootstrap logging/metrics.

`Identity` and its `Protected` holder move next to this use site under
Core/Backend, and `ProtectedTests` follows them.
