+++ Sources/Scout/Core/Backend/ActionTable.swift
@@ -36,26 +36,6 @@ private func run(_ action: ActionTable.Action) async {
     }
 }
 
-extension ActionTable {
-    static func appState(identity: Identity) -> ActionTable {
-        ActionTable(actions: [
-            AppLifecycle.willEnterForeground: {
-                try await persistentContainer.performBackgroundTasks(
-                    { try SessionObject.trigger(session: identity.session, launchID: identity.launch, in: $0) },
-                    { try UserActivityObject.trigger(sessionID: identity.session.current, in: $0) },
-                    { try VersionMarker.trigger(installID: identity.install, in: $0) }
-                )
-            },
-            AppLifecycle.didEnterBackground: {
-                try await persistentContainer.performBackgroundTasks(
-                    { try SessionObject.complete(launchID: identity.launch, in: $0) },
-                    { try UserActivityObject.trigger(sessionID: identity.session.current, in: $0) }
-                )
-            },
-        ])
-    }
-}
-
 enum AppLifecycle {
     static var willEnterForeground: Notification.Name {
         #if canImport(UIKit)
