+++ Tests/ScoutTests/Core/Backend/ProtectedTests.swift
